[Android] Kakao Map 추가

+++ app/src/main/java/com/eidem/integratedmovielookup/model/CGVTheaterData.kt
@@ -3,5 +3,5 @@ package com.eidem.integratedmovielookup.model
 data class CGVTheaterData(val theaterCd: String,
                           val theaterNm: String,
                           val areaCd: String,
-                          var lat: Float? = null,
-                          var lng: Float? = null)
+                          var lat: Double? = null,
+                          var lng: Double? = null)
